+++ CHANGELOG.md
@@ -37,6 +37,8 @@ Changelog
 - **Fix**: Stub hint function bodies if FIR-generated.
 - **Fix**: Hide FIR-generated hint functions.
 - **Fix**: Full support injected Java fields from Dagger-processed Java supertypes.
+- **Fix**: Use correct type args for empty map provider return types.
+- **Fix**: Add Dagger's internal `SetFactory`, `MapFactory` and `MapProviderFactory` to known Dagger ClassIds in `Provider` interop.
 - **Removed**: Removed deprecated `enableScopedInjectClassHints` Gradle DSL property.
 - **Removed**: Removed deprecated `enableStrictValidation` Gradle DSL property.
 - **Removed**: Removed deprecated `contributesGraphExtension` Gradle DSL property.
@@ -55,6 +57,7 @@ Special thanks to the following contributors for contributing to this release!
 - [@jonamireh](https://github.com/jonamireh)
 - [@zsmb13](https://github.com/zsmb13)
 - [@bjdodson](https://github.com/bjdodson)
+- [@consp1racy](https://github.com/consp1racy)
 
 0.7.7
 -----